app/src/main/java/com/adrpien/noteapp/feature_notes/data/data_source/local/NoteDao.kt
package com.adrpien.noteapp.feature_notes.data.data_source.local

import androidx.room.Dao
import androidx.room.Delete
import androidx.room.Insert
import androidx.room.OnConflictStrategy
import androidx.room.Query
import androidx.room.Update
import com.adrpien.noteapp.feature_notes.domain.model.Note
import kotlinx.coroutines.flow.Flow

@Dao
interface NoteDao {

    @Insert(onConflict = OnConflictStrategy.REPLACE)
    suspend fun insertNote(note: Note)

    @Delete
    suspend fun deleteNote(note: Note)

    @Query("SELECT * FROM note")
    fun getAllNotes(): Flow<List<Note>>

    // example query "SELECT SUM(price * amount) FROM table_name"
    @Query("SELECT * FROM note WHERE id LIKE :id")
    fun getNote(id: Int): Note

}